+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/Config.java
@@ -79,6 +79,8 @@ public interface Config {
 
     AutocrafterManagerEntry getAutocrafterManager();
 
+    SimpleEnergyUsageEntry getAutocraftingMonitor();
+
     interface SimpleEnergyUsageEntry {
         long getEnergyUsage();
     }