+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/content/Blocks.java
@@ -2,6 +2,7 @@ package com.refinedmods.refinedstorage.common.content;
 
 import com.refinedmods.refinedstorage.common.autocrafting.autocrafter.AutocrafterBlock;
 import com.refinedmods.refinedstorage.common.autocrafting.autocraftermanager.AutocrafterManagerBlock;
+import com.refinedmods.refinedstorage.common.autocrafting.monitor.AutocraftingMonitorBlock;
 import com.refinedmods.refinedstorage.common.autocrafting.patterngrid.PatternGridBlock;
 import com.refinedmods.refinedstorage.common.constructordestructor.AbstractConstructorBlockEntity;
 import com.refinedmods.refinedstorage.common.constructordestructor.AbstractDestructorBlockEntity;
@@ -156,6 +157,12 @@ public final class Blocks {
         AutocrafterManagerBlock::new,
         ContentIds.AUTOCRAFTER_MANAGER,
         ContentNames.AUTOCRAFTER_MANAGER,
+        COLOR
+    );
+    private final BlockColorMap<AutocraftingMonitorBlock, BaseBlockItem> autocraftingMonitor = new BlockColorMap<>(
+        AutocraftingMonitorBlock::new,
+        ContentIds.AUTOCRAFTING_MONITOR,
+        ContentNames.AUTOCRAFTING_MONITOR,
         COLOR
     );
 
@@ -432,5 +439,9 @@ public final class Blocks {
 
     public BlockColorMap<AutocrafterManagerBlock, BaseBlockItem> getAutocrafterManager() {
         return autocrafterManager;
+    }
+
+    public BlockColorMap<AutocraftingMonitorBlock, BaseBlockItem> getAutocraftingMonitor() {
+        return autocraftingMonitor;
     }
 }