+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/content/Tags.java
@@ -29,6 +29,7 @@ public final class Tags {
     public static final TagKey<Item> DISK_INTERFACES = createTag("disk_interfaces");
     public static final TagKey<Item> AUTOCRAFTERS = createTag("autocrafters");
     public static final TagKey<Item> AUTOCRAFTER_MANAGERS = createTag("autocrafter_managers");
+    public static final TagKey<Item> AUTOCRAFTING_MONITORS = createTag("autocrafting_monitors");
 
     private Tags() {
     }