+++ CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+-   Autocrafting Monitor
+
 ## [2.0.0-milestone.4.9] - 2024-11-01
 
 ### Added